refactor: Move public Checklist components to shared/BaseChecklist/Public
