+++ apps/editor.planx.uk/src/@planx/components/Checklist/Public.tsx
@@ -1,9 +1,9 @@
 import { PublicProps } from "@planx/components/shared/types";
 import React from "react";
 
-import type { ChecklistWithOptions } from "../model";
-import FlatChecklist from "./components/FlatChecklist";
-import GroupedChecklist from "./components/GroupedChecklist";
+import FlatChecklist from "../shared/BaseChecklist/Public/components/FlatChecklist"
+import GroupedChecklist from "../shared/BaseChecklist/Public/components/GroupedChecklist";
+import type { ChecklistWithOptions } from "./model";
 
 const ChecklistComponent: React.FC<PublicProps<ChecklistWithOptions>> = (
   props,
