+++ apps/editor.planx.uk/src/@planx/components/shared/BaseChecklist/Editor/components/Options/index.tsx
@@ -1,6 +1,6 @@
 import { ComponentType } from "@opensystemslab/planx-core/types";
-import { useCurrentOptions } from "@planx/components/Checklist/Public/hooks/useInitialOptions";
-import { AnyOption } from "@planx/components/Option/model";
+import type { AnyOption } from "@planx/components/Option/model";
+import { useCurrentOptions } from "@planx/components/shared/BaseChecklist/Public/hooks/useInitialOptions";
 import { getOptionsSchemaByFn } from "@planx/components/shared/utils";
 import { partition } from "lodash";
 import React from "react";
@@ -8,7 +8,7 @@ import { FormikHookReturn } from "types";
 import ListManager from "ui/editor/ListManager/ListManager";
 import ModalSectionContent from "ui/editor/ModalSectionContent";
 
-import { AnyChecklist } from "../../../model";
+import type { AnyChecklist } from "../../../model";
 import { ExclusiveOrOptionManager } from "./ExclusiveOrOptionManager";
 import { GroupedOptions } from "./GroupedOptions";
 import ChecklistOptionsEditor from "./OptionsEditor";
