+++ apps/editor.planx.uk/src/@planx/components/shared/BaseChecklist/Public/tests/Public.grouped.test.tsx
@@ -5,7 +5,7 @@ import { setup } from "testUtils";
 import { vi } from "vitest";
 import { axe } from "vitest-axe";
 
-import Checklist from "../Public";
+import Checklist from "../../../../Checklist/Public";
 import {
   groupedOptions,
   groupedOptionsWithExclusiveOption,
